refactor[menu]: menu와 menugroup 다른 애그리거트 분리
