+++ src/main/java/kitchenpos/menus/tobe/menu/application/MenuProductClient.java
@@ -1,6 +1,6 @@
-package kitchenpos.menus.tobe.application;
+package kitchenpos.menus.tobe.menu.application;
 
-import kitchenpos.menus.tobe.domain.ProductVO;
+import kitchenpos.menus.tobe.menu.ui.dto.ProductResponse;
 import kitchenpos.products.tobe.domain.Product;
 
 import java.util.List;
@@ -9,5 +9,5 @@ import java.util.UUID;
 public interface MenuProductClient {
     List<Product> findAllByIdIn(List<UUID> productIds);
     Product findById(UUID productId);
-    List<ProductVO> findAllByIdn(List<UUID> productIds);
+    List<ProductResponse> findAllByIdn(List<UUID> productIds);
 }
