+++ src/main/java/kitchenpos/menus/tobe/menu/domain/MenuProductLoader.java
@@ -1,5 +1,6 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.menu.domain;
 
+import kitchenpos.menus.tobe.menu.ui.dto.ProductResponse;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -7,19 +8,19 @@ import java.util.Objects;
 
 @Component
 public class MenuProductLoader {
-    public MenuProducts loadMenuProducts(final List<MenuProduct> menuProducts, final List<ProductVO> products) {
+    public MenuProducts loadMenuProducts(final List<MenuProduct> menuProducts, final List<ProductResponse> products) {
         validateRegistered(menuProducts, products);
         products.forEach(product -> loadProduct(menuProducts, product));
         return new MenuProducts(menuProducts);
     }
 
-    private void validateRegistered(final List<MenuProduct> menuProducts, final List<ProductVO> products) {
+    private void validateRegistered(final List<MenuProduct> menuProducts, final List<ProductResponse> products) {
         if (products.size() != menuProducts.size()) {
             throw new IllegalArgumentException("상품이 등록되어있지 않습니다.");
         }
     }
 
-    private void loadProduct(final List<MenuProduct> menuProducts, final ProductVO product) {
+    private void loadProduct(final List<MenuProduct> menuProducts, final ProductResponse product) {
         menuProducts.stream()
                 .filter(menuProduct -> Objects.equals(menuProduct.getProductId(), product.getId()))
                 .forEach(menuProduct -> menuProduct.loadPrice(product));
