+++ src/main/java/kitchenpos/menus/tobe/menu/domain/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.menu.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
