+++ src/main/java/kitchenpos/menus/tobe/menu/infra/RestMenuProductClient.java
@@ -1,14 +1,13 @@
-package kitchenpos.menus.tobe.infra;
+package kitchenpos.menus.tobe.menu.infra;
 
-import kitchenpos.menus.tobe.application.MenuProductClient;
-import kitchenpos.menus.tobe.domain.ProductVO;
+import kitchenpos.menus.tobe.menu.application.MenuProductClient;
+import kitchenpos.menus.tobe.menu.ui.dto.ProductResponse;
 import kitchenpos.products.tobe.application.ProductService;
 import kitchenpos.products.tobe.domain.Product;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 import static java.util.stream.Collectors.toList;
 
@@ -32,10 +31,10 @@ public class RestMenuProductClient implements MenuProductClient {
     }
 
     @Override
-    public List<ProductVO> findAllByIdn(List<UUID> productIds) {
+    public List<ProductResponse> findAllByIdn(List<UUID> productIds) {
         return productService.findAllByIdIn(productIds)
                 .stream()
-                .map(product -> new ProductVO(product.getId(), product.getPrice()))
+                .map(product -> new ProductResponse(product.getId(), product.getPrice()))
                 .collect(toList());
     }
 }
