+++ src/main/java/kitchenpos/menus/tobe/menugroup/domain/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.menugroup.domain;
 
 import java.util.List;
 import java.util.Optional;
