+++ src/main/java/kitchenpos/menus/tobe/menugroup/infra/JpaMenuGroupRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.menus.tobe.infra;
+package kitchenpos.menus.tobe.menugroup.infra;
 
-import kitchenpos.menus.tobe.domain.MenuGroup;
-import kitchenpos.menus.tobe.domain.MenuGroupRepository;
+import kitchenpos.menus.tobe.menugroup.domain.MenuGroup;
+import kitchenpos.menus.tobe.menugroup.domain.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
