+++ src/test/java/kitchenpos/menus/tobe/application/MenuServiceTest.java
@@ -5,13 +5,20 @@ import kitchenpos.common.event.FakeEventPublisher;
 import kitchenpos.common.infra.FakeProfanities;
 import kitchenpos.fixture.MenuFixture;
 import kitchenpos.fixture.ProductFixture;
-import kitchenpos.menus.tobe.domain.*;
 import kitchenpos.menus.tobe.infra.InMemoryMenuGroupRepository;
 import kitchenpos.menus.tobe.infra.InMemoryMenuRepository;
-import kitchenpos.menus.tobe.infra.RestMenuProductClient;
-import kitchenpos.menus.tobe.ui.dto.MenuChangePriceRequest;
-import kitchenpos.menus.tobe.ui.dto.MenuCreateRequest;
-import kitchenpos.menus.tobe.ui.dto.MenuProductRequest;
+import kitchenpos.menus.tobe.menu.infra.RestMenuProductClient;
+import kitchenpos.menus.tobe.menu.application.MenuService;
+import kitchenpos.menus.tobe.menu.domain.Menu;
+import kitchenpos.menus.tobe.menu.domain.MenuProductLoader;
+import kitchenpos.menus.tobe.menu.domain.MenuProducts;
+import kitchenpos.menus.tobe.menu.domain.MenuRepository;
+import kitchenpos.menus.tobe.menugroup.application.MenuGroupService;
+import kitchenpos.menus.tobe.menugroup.domain.MenuGroup;
+import kitchenpos.menus.tobe.menugroup.domain.MenuGroupRepository;
+import kitchenpos.menus.tobe.menu.ui.dto.MenuChangePriceRequest;
+import kitchenpos.menus.tobe.menu.ui.dto.MenuCreateRequest;
+import kitchenpos.menus.tobe.menu.ui.dto.MenuProductRequest;
 import kitchenpos.products.tobe.application.ProductService;
 import kitchenpos.products.tobe.infra.InMemoryProductRepository;
 import kitchenpos.products.tobe.domain.Product;
