+++ src/test/java/kitchenpos/menus/tobe/domain/MenuGroupTest.java
@@ -1,12 +1,11 @@
 package kitchenpos.menus.tobe.domain;
 
-import kitchenpos.fixture.MenuFixture;
+import kitchenpos.menus.tobe.menugroup.domain.MenuGroup;
 import org.assertj.core.api.ThrowableAssert;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.UUID;
