+++ src/test/java/kitchenpos/menus/tobe/domain/MenuProductTest.java
@@ -1,6 +1,9 @@
 package kitchenpos.menus.tobe.domain;
 
 import kitchenpos.fixture.MenuFixture;
+import kitchenpos.menus.tobe.menu.domain.MenuProduct;
+import kitchenpos.menus.tobe.menu.ui.dto.ProductResponse;
+import kitchenpos.menus.tobe.menu.domain.Quantity;
 import kitchenpos.products.tobe.domain.Price;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -27,7 +30,7 @@ class MenuProductTest {
     @Test
     @DisplayName("상품으로부터 가격을 설정하고 계산할 수 있다.")
     void loadProduct() {
-        final ProductVO product = new ProductVO(randomUUID(), new Price(BigDecimal.valueOf(1000L)));
+        final ProductResponse product = new ProductResponse(randomUUID(), new Price(BigDecimal.valueOf(1000L)));
         final long quantity = 2L;
         final MenuProduct menuProduct = MenuFixture.메뉴상품(product.getId(), quantity);
