+++ src/test/java/kitchenpos/menus/tobe/domain/QuantityTest.java
@@ -1,11 +1,11 @@
 package kitchenpos.menus.tobe.domain;
 
+import kitchenpos.menus.tobe.menu.domain.Quantity;
 import org.assertj.core.api.ThrowableAssert;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.springframework.beans.factory.annotation.Value;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
